+++ src/components/board.js
@@ -1,0 +1,5 @@
+export const createBoardTemplate = () => {
+  return `<section class="board container">
+  <div class="board__tasks"></div>
+  </section>`;
+};
