+++ src/components/loud-more.js
@@ -1,0 +1,3 @@
+export const createLoadMoreButtonTemplate = () => {
+  return `<button class="load-more" type="button">load more</button>`;
+};
